requirements.txt
fastapi
uvicorn[standard]
supertokens-python
asyncpg 
pydantic
pydantic-settings
pydantic[email]
pytest
sse-starlette
broadcaster
